+++ server/src/routes/student/deleteStudent.ts
@@ -1,4 +1,5 @@
 import express, {Request, Response} from 'express';
+import { BadRequestError } from '../../errors';
 import {requireAuth} from '../../middlewares';
 
 import { Student } from '../../models/students';
@@ -9,12 +10,15 @@ router.delete("/student/:id",
   requireAuth,
   async (req: Request, res: Response) => {
 
-    try {
-      await Student.findOneAndRemove({_id : req.params.id});
-      res.status(200).send({"message": "Student Deleted Successfully"});
-    } catch (err) {
-      console.log(err)
+    const student = await Student.findById(req.params.id);
+    
+    if(student){
+      if(student.userId !== req.currentUser!.id) throw new BadRequestError("Cannot delete the student of another mentor");
     }
+
+    await Student.findOneAndRemove({$and : [{"userId": req.currentUser!.id}, {_id : req.params.id}]});
+    res.status(200).send({"message": "Student Deleted Successfully"});
+  
     
   });
